# core-service/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.RoomAvatar:
    properties:
      created_at:
        type: string
      id:
        type: integer
      is_current:
        type: boolean
      uploaded_by:
        type: integer
    type: object
...
  httpapi.ErrorResponse:
    properties:
      error:
        example: room not found
        type: string
    type: object
...
info:
  contact: {}
  description: 'Core service of BuddyGym: auth, users, rooms, rewards, checkin proxy.'
  title: BuddyGym Core API
  version: "1.0"
paths:
...
  /rooms/{id}/avatar:
    get:
      description: Streams the newest picture of the room from private object storage.
        Requires a Bearer token, so browsers must fetch it via XHR rather than a plain
        <img src>. Returns 404 when the room has no picture.
      parameters:
      - description: room id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - image/jpeg
      responses:
        "200":
          description: OK
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Download the current room picture
      tags:
      - rooms
    put:
      consumes:
      - multipart/form-data
      description: Any room member may add one. Send multipart/form-data with a "photo"
        file up to 5 MB. The new picture becomes the face of the room; the previous
        ones stay in the gallery.
      parameters:
      - description: room id
        in: path
        name: id
        required: true
        type: integer
      - description: room picture
        in: formData
        name: photo
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.RoomAvatar'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Add a room picture
      tags:
      - rooms
  /rooms/{id}/avatars:
    get:
      description: Room members only. Newest first; the first entry is the picture
        the room currently wears.
      parameters:
      - description: room id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.RoomAvatar'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List the room picture gallery
      tags:
      - rooms
  /rooms/{id}/avatars/{avatarId}:
    delete:
      description: The member who uploaded it or the room creator. When the current
        picture goes, the room falls back to the newest one left.
      parameters:
      - description: room id
        in: path
        name: id
        required: true
        type: integer
      - description: picture id
        in: path
        name: avatarId
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a room picture
      tags:
      - rooms
    get:
      description: Room members only. Serves an older picture the room used to wear.
      parameters:
      - description: room id
        in: path
        name: id
        required: true
        type: integer
      - description: picture id
        in: path
        name: avatarId
        required: true
        type: integer
      produces:
      - image/jpeg
      responses:
        "200":
          description: OK
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpapi.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Download one picture from the gallery
      tags:
      - rooms
...
securityDefinitions:
  BearerAuth:
    description: 'JWT from POST /auth/telegram: "Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"